Collision hit box updated. also initial commit :)
